+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/rqes/collection_entries/Document.kt
@@ -3,6 +3,8 @@ package at.asitplus.dif.rqes.collection_entries
 import at.asitplus.dif.rqes.enums.ConformanceLevelEnum
 import at.asitplus.dif.rqes.enums.SignatureFormat
 import at.asitplus.dif.rqes.enums.SignedEnvelopeProperty
+import at.asitplus.dif.rqes.getSignAlgorithm
+import at.asitplus.signum.indispensable.Digest
 import at.asitplus.dif.rqes.serializers.Asn1EncodableBase64Serializer
 import at.asitplus.dif.rqes.getSignAlgorithm
 import at.asitplus.signum.indispensable.SignatureAlgorithm