+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/rqes/serializers/CSCSignatureRequestParameterSerializer.kt
@@ -1,14 +1,14 @@
 package at.asitplus.dif.rqes.serializers
 
+import at.asitplus.dif.rqes.CSCSignatureRequestParameters
 import at.asitplus.dif.rqes.SignDocParameters
 import at.asitplus.dif.rqes.SignHashParameters
-import at.asitplus.dif.rqes.SignatureRequestParameters
 import kotlinx.serialization.json.JsonContentPolymorphicSerializer
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.jsonObject
 
-object SignatureRequestParameterSerializer :
-    JsonContentPolymorphicSerializer<SignatureRequestParameters>(SignatureRequestParameters::class) {
+object CSCSignatureRequestParameterSerializer :
+    JsonContentPolymorphicSerializer<CSCSignatureRequestParameters>(CSCSignatureRequestParameters::class) {
     override fun selectDeserializer(element: JsonElement) = when {
         "hashes" in element.jsonObject -> SignHashParameters.serializer()
         else -> SignDocParameters.serializer()