+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/helper/RequestParser.kt
@@ -21,7 +21,7 @@ import io.ktor.util.*
 internal class AuthenticationRequestParser(
     /**
      * Need to implement if resources are defined by reference, i.e. the URL for a [JsonWebKeySet],
-     * or the authentication request itself as `request_uri`, or `presentation_definition_uri`.
+     * or the request itself as `request_uri`, or `presentation_definition_uri`.
      * Implementations need to fetch the url passed in, and return either the body, if there is one,
      * or the HTTP header `Location`, i.e. if the server sends the request object as a redirect.
      */
@@ -36,18 +36,18 @@ internal class AuthenticationRequestParser(
         fun createWithDefaults(
             remoteResourceRetriever: RemoteResourceRetrieverFunction? = null,
             requestObjectJwsVerifier: RequestObjectJwsVerifier? = null,
-        ) = AuthenticationRequestParser(
+        ) = RequestParser(
             remoteResourceRetriever = remoteResourceRetriever ?: { null },
             requestObjectJwsVerifier = requestObjectJwsVerifier ?: RequestObjectJwsVerifier { _, _ -> true },
         )
     }
 
     /**
-     * Pass in the URL sent by the Verifier (containing the [AuthenticationRequestParameters] as query parameters),
+     * Pass in the URL sent by the Verifier (containing the [RequestParameters] as query parameters),
      * to create [AuthenticationResponseParameters] that can be sent back to the Verifier, see
      * [AuthenticationResponseResult].
      */
-    suspend fun parseAuthenticationRequestParameters(input: String): KmmResult<AuthenticationRequestParametersFrom> = catching {
+    suspend fun parseRequestParameters(input: String): KmmResult<RequestParametersFrom> = catching {
         // maybe it is a request JWS
         val parsedParams = kotlin.run { parseRequestObjectJws(input) }
             ?: kotlin.runCatching { // maybe it's in the URL parameters
@@ -64,20 +64,21 @@ internal class AuthenticationRequestParser(
             ?: throw OAuth2Exception(Errors.INVALID_REQUEST)
                 .also { Napier.w("Could not parse authentication request: $input") }
 
-        val extractedParams = parsedParams.let { extractRequestObject(it.parameters) ?: it }
-            .also { Napier.i("Parsed authentication request: $it") }
+        val extractedParams =
+            parsedParams.let { extractRequestObject((it as AuthenticationRequestParametersFrom).parameters) ?: it }
+                .also { Napier.i("Parsed authentication request: $it") }
         extractedParams
     }
 
-    private suspend fun extractRequestObject(params: AuthenticationRequestParameters): AuthenticationRequestParametersFrom? =
+    private suspend fun extractRequestObject(params: AuthenticationRequestParameters): RequestParametersFrom? =
         params.request?.let { requestObject ->
             parseRequestObjectJws(requestObject)
         } ?: params.requestUri?.let { uri ->
             remoteResourceRetriever.invoke(uri)
-                ?.let { parseAuthenticationRequestParameters(it).getOrNull() }
+                ?.let { parseRequestParameters(it).getOrNull() }
         }
 
-    private fun parseRequestObjectJws(requestObject: String): AuthenticationRequestParametersFrom.JwsSigned? {
+    private fun parseRequestObjectJws(requestObject: String): RequestParametersFrom? {
         return JwsSigned.deserialize(requestObject).getOrNull()?.let { jws ->
             val params = AuthenticationRequestParameters.deserialize(jws.payload.decodeToString()).getOrElse {
                 return null